requirements.txt
Flask==2.0.2
Flask-OAuthlib==0.9.6
google-api-python-client==2.26.1
google-auth==2.3.0
google-auth-oauthlib==0.4.6
google-auth-httplib2==0.1.0
msal==1.14.0
APScheduler==3.8.0
Werkzeug==2.0.0  # Updated to version 2.0.0
WTForms==2.3.3
twilio==7.0.0
Flask-WTF==0.15.1  # Add this line for Flask-WTF dependency

requests~=2.31.0